pyproject.toml
[project]
name = "mymagicmirror"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash[async]>=3.2.0",
    "dash-mantine-components>=2.1.0",
    "google-api-python-client>=2.178.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.3",
    "dash-iconify>=0.1.2",
    "scapy>=2.5.0", # optional: active ARP scanning for Presence
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
    "pytest>=8.3.0",
]
